@@ -1,19 +1,18 @@
-'use strict';
+'use strict'
 
-const router = require('express-promise-router')();
+import {Request, Response} from 'express'
+import Router from 'express-promise-router'
+const router = Router()
 
-router.get('/', async function(req, res) {
-  await new Promise(function (resolve, reject) {
-    setTimeout(()=>resolve(), 2000);
-  });
-  res.json({message: 'await'});
-});
+router.get('/', async (req, res) => {
+  await new Promise( (resolve, reject) => {
+    setTimeout(() => resolve(), 2000)
+  })
+  res.json({message: 'await'})
+})
 
-router.get('/error', async function (req, res) {
-  let err = new Error('some thing happened');
-  err['req'] = req.url;
-  throw err;
-});
+router.get('/errors', async (req: Request, res: Response) => {
+  throw new Error('some error')
+})
 
-
-module.exports = router;
+module.exports = router
