@@ -1,9 +1,9 @@
 /* eslint-disable */
-require('eventsource-polyfill');
-const hotClient = require('webpack-hot-middleware/client?noInfo=true&reload=true');
+require('eventsource-polyfill')
+const hotClient = require('webpack-hot-middleware/client?noInfo=true&reload=true')
 
-hotClient.subscribe(function (event) {
+hotClient.subscribe((event) => {
   if (event.action === 'reload') {
     window.location.reload()
   }
-});
+})
