@@ -1,7 +1,5 @@
-'use strict';
-const express = require('express');
-const router = express.Router();
+'use strict'
+const express = require('express')
+const router  = express.Router()
 
-
-
-module.exports = router;
+module.exports = router
